Flytt typar brukt i brukermodellen inn til brukermodellen slik at alle brukardata-typar er samla.
Neste steg: Lage eigne typar for filtermodellen i staden for å gjenbruke desse.
