+++ src/components/fargekategori/fargekategori-popover.tsx
@@ -5,7 +5,8 @@ import {ThunkDispatch} from 'redux-thunk';
 import {Button, Popover} from '@navikt/ds-react';
 import {AppState} from '../../reducer';
 import {FARGEKATEGORI_OPPDATER_OK, oppdaterFargekategoriAction} from '../../ducks/fargekategori';
-import {FargekategoriDataModell, FargekategoriModell, Fargekategorinavn} from '../../model-interfaces';
+import {FargekategoriDataModell, Fargekategorinavn} from '../../model-interfaces';
+import {FargekategoriModell} from '../../typer/bruker-modell';
 import {fargekategoriIkonMapper} from './fargekategori-ikon-mapper';
 import {FargekategoriFeilhandtering} from './FargekategoriFeilhandtering';
 import {hentStatustallForVeileder, leggTilStatustall} from '../../ducks/statustall/statustall-veileder';
